rework test case fix bug

+++ test/test_pips.py
@@ -7,36 +7,47 @@ from unittest.mock import patch
 
 
 class PipsTest(unittest.TestCase):
+    def setUp(self) -> None:
+        #self.path_to_site_packages = "../venv36/Lib/site-packages/"
+        # on MacOS
+        self.path_to_site_packages = "../venv36/lib/python3.6/site-packages/"
+        self.package = "Jinja2"
+        self.sub_dependency = "MarkupSafe"
+        self.requirements_file = "requirements.txt"
+
     def test_install(self):
         """Tests if all packages from the requirements.txt are installed"""
-
-        if not os.path.isfile("requirements.txt"):
-            f = open("requirements.txt", "w+")
+        print(os.getcwd())
+        if not os.path.isfile(self.requirements_file):
+            f = open(self.requirements_file, "w+")
             f.close()
-        with open("requirements.txt", "w") as f:
-            f.writelines("jinja2")
+        with open(self.requirements_file, "w") as f:
+            f.writelines(self.package)
         test_args = ["pips", "install"]
+
         with patch.object(sys, 'argv', test_args):
             Pips()
+
+        self.assertTrue(os.path.exists(self.path_to_site_packages + self.package.lower()))
+        self.assertTrue(os.path.exists(self.requirements_file))
+        self.assertTrue(os.path.exists("requirements.lock"))
 
     def test_install_package(self):
         """Tests if a single package can be installed and locked"""
 
-        package = "Jinja2"
-        sub_dependency = "MarkupSafe"
-        test_args = ["pips", "install", package]
+        test_args = ["pips", "install", self.package]
         with patch.object(sys, 'argv', test_args):
             Pips()
 
-        self.assertTrue(os.path.exists("../venv37/Lib/site-packages/" + package.lower()))
-        self.assertTrue(os.path.exists("requirements.txt"))
+        self.assertTrue(os.path.exists(self.path_to_site_packages + self.package.lower()))
+        self.assertTrue(os.path.exists(self.requirements_file))
         self.assertTrue(os.path.exists("requirements.lock"))
 
         with open("requirements.txt", "r") as f:
             lines = f.readlines()
             package_found = False
             for line in lines:
-                if package in line:
+                if self.package in line:
                     package_found = True
             self.assertTrue(package_found)
 
@@ -45,30 +56,30 @@ class PipsTest(unittest.TestCase):
             package_found = False
             sub_dep_found = False
             for line in lines:
-                if package in line:
+                if self.package in line:
                     package_found = True
-                if sub_dependency in line:
+                if self.sub_dependency in line:
                     sub_dep_found = True
             self.assertTrue(package_found)
             self.assertTrue(sub_dep_found)
 
     def test_uninstall_package(self):
         """Tests if a single package can be uninstalled"""
-        package = "Jinja2"
-        sub_dependency = "MarkupSafe"
-        test_args = ["pips", "install", package]
+
+        test_args = ["pips", "install", self.package]
         with patch.object(sys, 'argv', test_args):
             Pips()
 
-        test_args = ["pips", "uninstall", package]
+        test_args = ["pips", "uninstall", self.package]
         with patch.object(sys, 'argv', test_args):
             Pips()
 
-        self.assertFalse(os.path.exists("../venv37/Lib/site-packages/" + package.lower()))
-        self.assertFalse(os.path.exists("../venv37/Lib/site-packages/" + sub_dependency.lower()))
+        self.assertFalse(os.path.exists(self.path_to_site_packages + self.package.lower()))
+        self.assertFalse(os.path.exists(self.path_to_site_packages + self.sub_dependency.lower()))
 
     def tearDown(self) -> None:
-        process = subprocess.Popen("pip uninstall --yes flask", shell=True,
+        print("teardown")
+        process = subprocess.Popen("pip uninstall --yes" + self.package, shell=True,
                                    stdout=subprocess.PIPE,
                                    stderr=subprocess.PIPE)
 
@@ -76,8 +87,10 @@ class PipsTest(unittest.TestCase):
         out, err = process.communicate()
         errcode = process.returncode
         print(out)
-        os.remove("requirements.txt")
-        os.remove("requirements.lock")
+        if os.path.exists(self.requirements_file):
+            os.remove(self.requirements_file)
+        if os.path.exists("requirements.lock"):
+            os.remove("requirements.lock")
 
 
 if __name__ == '__main__':
